refactor(docs): slider dx
