+++ apps/docs/content/components/slider/controlled-range.ts
@@ -1,50 +1,5 @@
-const App = `import {Slider} from "@nextui-org/react";
-
-export default function App() {
-  const [value, setValue] = React.useState([100, 300]);
-
-  return (
-    <div className="flex flex-col gap-2 w-full h-full max-w-md items-start justify-center">
-      <Slider 
-        label="Select a budget"
-        formatOptions={{style: "currency", currency: "USD"}}
-        step={10}
-        maxValue={1000}
-        minValue={0}
-        value={value} 
-        onChange={setValue}
-        className="max-w-md"
-      />
-      <p className="text-default-500 font-medium text-small">
-        Selected budget: {Array.isArray(value) && value.map((b) => \`$\${b}\`).join(" – ")}
-      </p>
-    </div>
-  );
-}`;
-
-const AppTs = `import {Slider, SliderValue} from "@nextui-org/react";
-
-export default function App() {
-  const [value, setValue] = React.useState<SliderValue>([100, 300]);
-
-  return (
-    <div className="flex flex-col gap-2 w-full h-full max-w-md items-start justify-center">
-      <Slider 
-        label="Select a budget"
-        formatOptions={{style: "currency", currency: "USD"}}
-        step={10}
-        maxValue={1000}
-        minValue={0}
-        value={value} 
-        onChange={setValue}
-        className="max-w-md"
-      />
-      <p className="text-default-500 font-medium text-small">
-        Selected budget: {Array.isArray(value) && value.map((b) => \`$\${b}\`).join(" – ")}
-      </p>
-    </div>
-  );
-}`;
+import App from "./controlled-range.raw.jsx?raw";
+import AppTs from "./controlled-range.raw.tsx?raw";
 
 const react = {
   "/App.jsx": App,
