+++ apps/docs/content/components/slider/fill-offset.ts
@@ -1,21 +1,4 @@
-const App = `import {Slider} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Slider 
-      label="Exposure" 
-      size="sm"
-      color="warning"
-      step={0.01} 
-      maxValue={5} 
-      minValue={-5} 
-      fillOffset={0}
-      defaultValue={1.5}
-      className="max-w-md"
-      formatOptions={{signDisplay: 'always'}}
-    />
-  );
-}`;
+import App from "./fill-offset.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
