+++ apps/docs/content/components/slider/get-value.ts
@@ -1,16 +1,4 @@
-const App = `import {Slider} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Slider 
-      label="Donuts to buy"
-      size="sm"
-      maxValue={60}
-      getValue={(donuts) => \`\${donuts} of 60 Donuts\`}
-      className="max-w-md"
-    />
-  );
-}`;
+import App from "./get-value.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
