+++ apps/docs/content/components/slider/hide-thumb.ts
@@ -1,16 +1,4 @@
-const App = `import {Slider} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Slider 
-      aria-label="Player progress" 
-      color="foreground"
-      hideThumb={true}
-      defaultValue={20}
-      className="max-w-md"
-    />
-  );
-}`;
+import App from "./hide-thumb.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
