+++ apps/docs/content/components/slider/value-formatting.ts
@@ -1,17 +1,4 @@
-const App = `import {Slider} from "@nextui-org/react";
-
-export default function App() {
-  return (
-    <Slider 
-      label="Currency" 
-      showTooltip={true}
-      formatOptions={{style: 'currency', currency: 'JPY'}}
-      tooltipValueFormatOptions={{style: 'currency', currency: 'JPY'}}
-      defaultValue={40}
-      className="max-w-md"
-    />
-  );
-}`;
+import App from "./value-formatting.raw.jsx?raw";
 
 const react = {
   "/App.jsx": App,
